# docs/swagger.yaml
basePath: /api/
definitions:
...
  httputil.HTTPError:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  models.CreatePlayer:
    properties:
      age:
        example: 25
        type: integer
      country:
        example: Germany
        type: string
      first_name:
        example: Audrey
        type: string
      last_name:
        example: Hepburn
        type: string
      market_value:
        example: 25000
        type: integer
      position:
        example: 1
        type: integer
    type: object
...
  models.ShowPlayer:
    properties:
      age:
        example: 25
        type: integer
      country:
        example: Germany
        type: string
      first_name:
        example: Audrey
        type: string
      id:
        type: integer
      last_name:
        example: Hepburn
        type: string
      market_value:
        example: 25000
        type: integer
      position:
        example: 1
        type: integer
    type: object
  models.ShowTeam:
    properties:
      budget:
        type: integer
      country:
        type: string
      id:
        type: integer
      market_value:
        type: integer
      name:
        type: string
      players:
        items:
          $ref: '#/definitions/models.ShowPlayer'
        type: array
    type: object
...
  models.UpdatePlayer:
    properties:
      age:
        example: 25
        type: integer
      country:
        example: Germany
        type: string
      first_name:
        example: Audrey
        type: string
      last_name:
        example: Hepburn
        type: string
      market_value:
        example: 25000
        type: integer
      position:
        example: 1
        type: integer
      team:
        type: integer
    type: object
  models.UpdateTeam:
    properties:
      budget:
        type: integer
      country:
        type: string
      name:
        type: string
      owner:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Fantasy football manager microservice.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Fantasy football manager API
  version: "1.0"
paths:
...
  /teams:
    post:
      consumes:
      - application/json
      description: Create a new team
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      security:
      - BearerAuth: []
      summary: Post a team
      tags:
      - Teams
  /teams/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a team and all of it's players
      parameters:
      - description: Team ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      security:
      - BearerAuth: []
      summary: Delete a team and all of it's players
      tags:
      - Teams
    get:
      consumes:
      - application/json
      description: Get team by ID
      parameters:
      - description: Team ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ShowTeam'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Get a team
      tags:
      - Teams
    patch:
      consumes:
      - application/json
      description: Update a team
      parameters:
      - description: Team ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update team payload
        in: body
        name: team
        required: true
        schema:
          $ref: '#/definitions/models.UpdateTeam'
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      security:
      - BearerAuth: []
      summary: Update a team
      tags:
      - Teams
  /teams/{id}/players:
    get:
      consumes:
      - application/json
      description: List all the players of a team
      parameters:
      - description: Team ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ShowPlayer'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: List players of a team
      tags:
      - Teams
    post:
      consumes:
      - application/json
      description: Create a player on a new team
      parameters:
      - description: Create player
        in: body
        name: player
        required: true
        schema:
          $ref: '#/definitions/models.CreatePlayer'
      - description: Team ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      security:
      - BearerAuth: []
      summary: Create a player on a new team
      tags:
      - Teams
  /teams/{teamId}/players/{id}:
    get:
      consumes:
      - application/json
      description: Get a player by ID from a team.
      parameters:
      - description: Player ID
        in: path
        name: id
        required: true
        type: integer
      - description: Team ID
        in: path
        name: teamId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ShowPlayer'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Show a player from a team
      tags:
      - Teams
    patch:
      consumes:
      - application/json
      description: Update a player from a team
      parameters:
      - description: Update player
        in: body
        name: player
        required: true
        schema:
          $ref: '#/definitions/models.UpdatePlayer'
      - description: Player ID
        in: path
        name: id
        required: true
        type: integer
      - description: Team ID
        in: path
        name: teamId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      security:
      - BearerAuth: []
      summary: Update a player from a team
      tags:
      - Teams
...
  /users/me/team/players:
    get:
      consumes:
      - application/json
      description: Get the logged in user's team players
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ShowPlayer'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      security:
      - BearerAuth: []
      summary: Get the logged in user's team players
      tags:
      - Users
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"